show

show method for CoverageExperiment and AggregatedCoverage objects

Usage

# S4 method for class 'CoverageExperiment'
show(object)

# S4 method for class 'AggregatedCoverage'
show(object)

# S3 method for class 'CoverageExperiment'
print(x, ..., n = NULL)

# S3 method for class 'AggregatedCoverage'
print(x, ..., n = NULL)

# S3 method for class 'tidyCoverageExperiment'
tbl_format_header(x, setup, ...)

# S3 method for class 'tidyAggregatedCoverage'
tbl_format_header(x, setup, ...)

Arguments

object

a CoverageExperiment or AggregatedCoverage object

x

Object to format or print.

...

Passed on to tbl_format_setup().

n

Number of rows to show. If NULL, the default, will print all rows if less than the print_max option. Otherwise, will print as many rows as specified by the print_min option.

setup

a setup object returned from pillar::tbl_format_setup().

Value

Prints a message to the console describing the contents of the CoverageExperimentorAggregatedCoverage` objects.

Examples

data(ce)
print(ce)
#> class: CoverageExperiment 
#> dim: 1 2 
#> metadata(0):
#> assays(1): coverage
#> rownames(1): Scc1
#> rowData names(2): features n
#> colnames(2): RNA_fwd RNA_rev
#> colData names(1): track
#> width: 3000
data(ac)
print(ac)
#> class: AggregatedCoverage 
#> dim: 1 2 
#> metadata(0):
#> assays(8): mean median ... ci_low ci_high
#> rownames(1): Scc1
#> rowData names(1): features
#> colnames(2): RNA_fwd RNA_rev
#> colData names(1): track
#> width: 3000
#> binning: 1