show

show method for CoverageExperiment and AggregatedCoverage objects

Usage

# S4 method for CoverageExperiment
show(object)

# S4 method for AggregatedCoverage
show(object)

# S3 method for AggregatedCoverage
print(x, ..., n = NULL)

# S3 method for tidyAggregatedCoverage
tbl_format_header(x, setup, ...)

Arguments

object

a CoverageExperiment or AggregatedCoverage object

x

Object to format or print.

...

Passed on to tbl_format_setup().

n

Number of rows to show. If NULL, the default, will print all rows if less than the print_max option. Otherwise, will print as many rows as specified by the print_min option.

setup

a setup object returned from pillar::tbl_format_setup().

Value

Prints a message to the console describing the contents of the CoverageExperimentorAggregatedCoverage` objects.

Examples

data(ce)
print(ce)
#> class: CoverageExperiment 
#> dim: 1 2 
#> metadata(0):
#> assays(1): coverage
#> rownames(1): Scc1
#> rowData names(1): features
#> colnames(2): RNA_fwd RNA_rev
#> colData names(1): track
#> width: 3000
data(ac)
print(ac)
#> # An AggregatedCoverage-tibble abstraction: 6000 × 11
#> # features=1 | tracks=2 | assays=mean, median, min, max, sd, se, ci_low,
#> #   ci_high
#> # width=3000 | binning=1
#>    track   features coord    mean median   min   max    sd     se ci_low ci_high
#>    <chr>   <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>   <dbl>
#>  1 RNA_fwd Scc1     -1500 -0.0921 -0.611 -3.25  10.5  1.42 0.0572 -0.204  0.0203
#>  2 RNA_fwd Scc1     -1499 -0.0915 -0.609 -3.25  10.5  1.42 0.0572 -0.204  0.0208
#>  3 RNA_fwd Scc1     -1498 -0.0898 -0.609 -3.25  10.5  1.42 0.0573 -0.202  0.0227
#>  4 RNA_fwd Scc1     -1497 -0.0914 -0.609 -3.25  10.5  1.42 0.0573 -0.204  0.0211
#>  5 RNA_fwd Scc1     -1496 -0.0915 -0.609 -3.25  10.5  1.42 0.0573 -0.204  0.0211
#>  6 RNA_fwd Scc1     -1495 -0.0912 -0.609 -3.25  10.5  1.42 0.0572 -0.204  0.0212
#>  7 RNA_fwd Scc1     -1494 -0.0912 -0.609 -3.25  10.5  1.42 0.0572 -0.204  0.0211
#>  8 RNA_fwd Scc1     -1493 -0.0915 -0.609 -3.25  10.5  1.42 0.0572 -0.204  0.0209
#>  9 RNA_fwd Scc1     -1492 -0.0907 -0.609 -3.25  10.5  1.42 0.0573 -0.203  0.0218
#> 10 RNA_fwd Scc1     -1491 -0.0903 -0.609 -3.25  10.5  1.42 0.0572 -0.203  0.0220
#> # ℹ 5,990 more rows